EntityManager added
EntityManager is added and a few other things are added/corrected mostly related to EntityManager or the entity system in general.

unrelated small change: Launcher class is removed, main method added to the Game class. This might not be the final location of the main method, note that the gameManager is not implemented yet and Game class itself is used for most of GameManager stuff. (It's not hard to seperate them later)

Note to self: move this directory to the Update branch or make a new branch for this.